requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asyncio==3.4.3
attrs==23.2.0
audioread==3.0.1
azure-cognitiveservices-speech==1.37.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.111.0
fastapi-cli==0.0.2
ffmpeg==1.4
filelock==3.14.0
flatbuffers==24.3.25
future==1.0.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanfriendly==10.0
idna==3.6
imageio==2.34.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain-core==0.1.50
langchain-openai==0.1.6
langsmith==0.1.54
lazy_loader==0.4
librosa==0.10.2
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
mypy==1.10.0
mypy-extensions==1.0.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
onnxruntime==1.17.3
openai==1.25.2
opencv-python-headless==4.9.0.80
orjson==3.10.3
packaging==23.2
pillow==10.3.0
platformdirs==4.2.1
pooch==1.8.1
protobuf==5.26.1
pycparser==2.22
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyMatting==1.1.12
pyreadline3==3.4.1
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.35.1
regex==2024.4.28
rembg==2.0.56
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
shellingham==1.5.4
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.5.0
tifffile==2024.5.3
tiktoken==0.6.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.1
watchfiles==0.21.0
websockets==12.0
